+++ src/Set.php
@@ -154,11 +154,17 @@ final class Set implements \Countable
         return $self;
     }
 
+    /**
+     * @return 0|positive-int
+     */
     public function size(): int
     {
         return $this->implementation->size();
     }
 
+    /**
+     * @return 0|positive-int
+     */
     public function count(): int
     {
         return $this->implementation->size();
@@ -237,6 +243,22 @@ final class Set implements \Countable
     }
 
     /**
+     * This is the same behaviour as `filter` but it allows Psalm to understand
+     * the type of the values contained in the returned Set
+     *
+     * @template S
+     *
+     * @param Predicate<S> $predicate
+     *
+     * @return self<S>
+     */
+    public function keep(Predicate $predicate): self
+    {
+        /** @var self<S> */
+        return $this->filter($predicate);
+    }
+
+    /**
      * Return all elements that satisfy the given predicate
      *
      * @param callable(T): bool $predicate
@@ -246,6 +268,19 @@ final class Set implements \Countable
     public function filter(callable $predicate): self
     {
         return new self($this->implementation->filter($predicate));
+    }
+
+    /**
+     * Return all elements that don't satisfy the given predicate
+     *
+     * @param callable(T): bool $predicate
+     *
+     * @return self<T>
+     */
+    public function exclude(callable $predicate): self
+    {
+        /** @psalm-suppress MixedArgument */
+        return $this->filter(static fn($value) => !$predicate($value));
     }
 
     /**